changed injector constant to all capital letters; renamed some classes
