+++ src/main/java/vlad/petrovskyi/internetshop/controllers/db/products/AddProductController.java
@@ -12,9 +12,9 @@ import vlad.petrovskyi.internetshop.service.ProductService;
 
 public class AddProductController extends HttpServlet {
 
-    private static final Injector injector = Injector.getInstance("vlad.petrovskyi.internetshop");
+    private static final Injector INJECTOR = Injector.getInstance("vlad.petrovskyi.internetshop");
     private final ProductService productService =
-            (ProductService) injector.getInstance(ProductService.class);
+            (ProductService) INJECTOR.getInstance(ProductService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
