+++ src/main/java/vlad/petrovskyi/internetshop/controllers/db/users/DeleteUserController.java
@@ -10,8 +10,8 @@ import vlad.petrovskyi.internetshop.service.UserService;
 
 public class DeleteUserController extends HttpServlet {
 
-    private static final Injector injector = Injector.getInstance("vlad.petrovskyi.internetshop");
-    private final UserService userService = (UserService) injector.getInstance(UserService.class);
+    private static final Injector INJECTOR = Injector.getInstance("vlad.petrovskyi.internetshop");
+    private final UserService userService = (UserService) INJECTOR.getInstance(UserService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
