+++ src/main/java/vlad/petrovskyi/internetshop/controllers/user/GetAllProductsController.java
@@ -1,4 +1,4 @@
-package vlad.petrovskyi.internetshop.controllers.db.products;
+package vlad.petrovskyi.internetshop.controllers.user;
 
 import java.io.IOException;
 import java.util.List;
@@ -10,18 +10,17 @@ import vlad.petrovskyi.internetshop.lib.Injector;
 import vlad.petrovskyi.internetshop.model.Product;
 import vlad.petrovskyi.internetshop.service.ProductService;
 
-public class AllProductsFromDbController extends HttpServlet {
+public class GetAllProductsController extends HttpServlet {
 
-    private static final Injector injector = Injector.getInstance("vlad.petrovskyi.internetshop");
+    private static final Injector INJECTOR = Injector.getInstance("vlad.petrovskyi.internetshop");
     private final ProductService productService =
-            (ProductService) injector.getInstance(ProductService.class);
+            (ProductService) INJECTOR.getInstance(ProductService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
         List<Product> productList = productService.getAll();
-
         req.setAttribute("products", productList);
-        req.getRequestDispatcher("/WEB-INF/views/products/allProductsEdit.jsp").forward(req, resp);
+        req.getRequestDispatcher("WEB-INF/views/products/allProductsBuy.jsp").forward(req, resp);
     }
 }
