+++ src/main/java/vlad/petrovskyi/internetshop/controllers/user/RegistrationController.java
@@ -12,10 +12,10 @@ import vlad.petrovskyi.internetshop.service.UserService;
 
 public class RegistrationController extends HttpServlet {
 
-    private static final Injector injector = Injector.getInstance("vlad.petrovskyi.internetshop");
-    private final UserService userService = (UserService) injector.getInstance(UserService.class);
+    private static final Injector INJECTOR = Injector.getInstance("vlad.petrovskyi.internetshop");
+    private final UserService userService = (UserService) INJECTOR.getInstance(UserService.class);
     private final ShoppingCartService shoppingCartService =
-            (ShoppingCartService) injector.getInstance(ShoppingCartService.class);
+            (ShoppingCartService) INJECTOR.getInstance(ShoppingCartService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
