+++ src/main/webapp/web.xml
@@ -23,7 +23,7 @@
 
     <servlet>
         <servlet-name>getAllUsers</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.users.AllUsersController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.users.GetAllUsersController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>getAllUsers</servlet-name>
@@ -41,7 +41,7 @@
 
     <servlet>
         <servlet-name>getAllProductsFromDb</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.products.AllProductsFromDbController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.products.GetAllProductsController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>getAllProductsFromDb</servlet-name>
@@ -59,7 +59,7 @@
 
     <servlet>
         <servlet-name>deleteProductFromDb</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.products.DeleteProductFromDbController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.products.DeleteProductController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>deleteProductFromDb</servlet-name>
@@ -77,7 +77,7 @@
 
     <servlet>
         <servlet-name>allProductsToBuy</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.user.AllProductsToBuyController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.user.GetAllProductsController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>allProductsToBuy</servlet-name>
@@ -86,7 +86,7 @@
 
     <servlet>
         <servlet-name>addToCart</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.shoppingcart.AddToCartController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.shoppingcart.AddProductToCartController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>addToCart</servlet-name>
